Llama2

Documentation for Llama2.

Llama2.Tokenizer

Llama2.Tokenizer

Llama2.Tokenizer — Type

Used for mapping from strings to token arrays (Int vectors) and back.

llama.c correspondence: Tokenizer (l. 372)

indextotoken = vocab

tokentoindex = sorted_vocab

removed maxtokenlength (not required in Julia)

removed byte_pieces (not required in Julia)

Llama2.Tokenizer — Method

Tokenizer(tokenizer_path::String, vocab_size::Int)

Constructs a Tokenizer by loading the vocabulary from a file in the llama2.c format. The vocabulary size must be known from the config.

llama.c correspondence: build_tokenizer (l. 385)