Llama2.jl

What is Llama2?

LLama2 is a family of pre-trained LLMs by Meta AI. More information can be found at: https://www.llama.com/

What is Llama2.jl?

Llama2.jl can inference a given model from within julia. For this cause you will have to provide your own model checkpoint. This project follows the procedure outlined by the run.c file from llama2.c.

Getting started

Clone the repository to a desired location:

cd /PATH/TO/DESIRED/LOCATION/
git clone git@github.com:ConstantConstantin/Llama2.git

Start julia, activate a desired environment and add the package; it can then be loaded in your session:

(@v1.11) pkg> activate .
  Activating new project at `PATH/TO/MY/ENVIRONMENT/myLlama2`

(myLlama2) pkg> add /PATH/TO/DESIRED/LOCATION/Llama2/
     Cloning git-repo `/PATH/TO/DESIRED/LOCATION/Llama2`
    Updating git-repo `/PATH/TO/DESIRED/LOCATION/Llama2`
    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
    Updating `PATH/TO/MY/ENVIRONMENT/Project.toml`
  [0e620e9f] + Llama2 v1.0.0-DEV `/PATH/TO/DESIRED/LOCATION/Llama2#aj/docs`
    Updating `PATH/TO/MY/ENVIRONMENT/Manifest.toml`
  [0e620e9f] + Llama2 v1.0.0-DEV `/PATH/TO/DESIRED/LOCATION/Llama2#aj/docs`
Precompiling project...
  1 dependency successfully precompiled in 1 seconds

julia> using Llama2

Llama2.TokenIndex

Llama2.Tokenizer

Llama2.compare_tokens

Llama2.TokenIndex — Type

TokenIndex(str::AbstractVector{UInt8}, id::Integer)

Create a TokenIndex from a byte vector and an integer identifier.

The byte sequence is converted to Vector{UInt8} and the ID is converted to Int16. Throw a DomainError if id ≤ 0.

Examples

julia> using Llama2;

julia> TokenIndex([0x61], 1)
TokenIndex(UInt8[0x61], 1)

julia> TokenIndex([0x61], -1)
ERROR: DomainError with Token index must be > 0.
[...]

Developer Notes

This is an internal struct.

Llama2.Tokenizer — Type

Tokenizer

Construct a tokenizer storing vocabulary entries, scores, and byte-piece mappings.

Constructors

Tokenizer(vocab, vocab_scores, sorted_vocab, vocab_size, max_token_length, byte_pieces) Construct a tokenizer directly from the provided fields. Validate that max_token_length > 0 and that byte_pieces has length 256.

Tokenizer(path::String, vocab_size::Integer) Load a tokenizer from a binary file.

Fields

vocab: Token byte sequences.

vocab_scores: Scores for each token.

sorted_vocab: Sorted token indices.

vocab_size: Number of vocabulary entries.

max_token_length: Maximum token length in bytes.

byte_pieces: Byte mapping (length 256).

Llama2.compare_tokens — Method

compare_tokens(first_token::TokenIndex, second_token::TokenIndex) -> Bool

Compare two TokenIndex objects for equality based solely on their byte-string contents. Return true if both tokens contain identical str fields, regardless of ID.

Examples

julia> using Llama2;

julia> compare_tokens(TokenIndex([0x61], 1), TokenIndex([0x61], 2))
true

julia> compare_tokens(TokenIndex([0x61], 1), TokenIndex([0x62], 1))
false