Llama2

Documentation for Llama2.

Llama2.Config

Llama2.RunState

Llama2.RunState

Llama2.Tokenizer

Llama2.Tokenizer

Llama2.Transformer

Llama2.TransformerWeights

Llama2.TransformerWeights

Llama2.Config — Type

Configuration of initial parameters

llama2.c correspondence Config (l.19)

Llama2.RunState — Type

Initial parameters for the run state

llama2.c correspondence: RunState (l. 50)

Llama2.RunState — Method

Initialization of RunState based on Config

initialized as undefined arrays

llama2.c correspondence: mallocrunstate (l. 77)

Llama2.Tokenizer — Type

Used for mapping from strings to token arrays (Int vectors) and back.

llama.c correspondence: Tokenizer (l. 372)

indextotoken = vocab

tokentoindex = sorted_vocab

removed maxtokenlength (not required in Julia)

removed byte_pieces (not required in Julia)

Llama2.Tokenizer — Method

Tokenizer(tokenizer_path::String, vocab_size::Int)

Constructs a Tokenizer by loading the vocabulary from a file in the llama2.c format. The vocabulary size must be known from the config.

llama.c correspondence: build_tokenizer (l. 385)

Llama2.Transformer — Type

Initial parameters for the transformer

llama2.c correspondence: Transformer (l. 67)

Llama2.TransformerWeights — Type

Initial parameters for the transformer weights

llama2.c correspondence: TransformerWeights (l. 29)

Llama2.TransformerWeights — Method

Initialize transformer weights based on Config

initialized as undefined arrays

llama2.c correspondence: memorymapweights (l. 111)